Add check to reload data on datasetgroup change
When updating the dataset group to a different group
that has the same period id the table did not reload
with the new mechanisms. When the dataset group is now
changed the period is checked if it has stayed the same
if it has then the table is updated.

References jqueipo/DATIM#663